Handling for percentile aggregate function (#822)
While data loading changed the data type to Long instead of Double

+++ integration/spark/src/main/scala/org/apache/spark/sql/CarbonDatasourceRelation.scala
@@ -229,7 +229,7 @@ case class CarbonRelation(
         .map(x => AttributeReference(x.getColName, CarbonMetastoreTypes.toDataType(
         metaData.carbonTable.getMeasureByName(factTable, x.getColName).getDataType.toString
           .toLowerCase match {
-          case "int" => "double"
+          case "int" => "long"
           case "decimal" => "decimal(" + x.getPrecision + "," + x.getScale + ")"
           case others => others
         }),
